requirements.dev.txt
fastapi==0.103.1
psycopg==3.1.10
pydantic-settings==2.0.3
SQLAlchemy==2.0.21
typer==0.9.0
uvicorn==0.23.2

pytest==7.4.2
pytest-asyncio==0.21.1